life-discovery/services/discovery-engine/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.4
pydantic-settings==2.7.0
sqlalchemy==2.0.36
psycopg[binary]==3.2.3
python-dateutil==2.9.0.post0
